+++ src/core.cpp
@@ -114,15 +114,18 @@ void raiseShellExit(ARState &st, int64_t status) {
 }
 
 void raiseAssertFail(ARState &st, Value &&msg, const AssertOp op, Value &&left, Value &&right) {
+  constexpr auto MAX_PRINTABLE = SYS_LIMIT_PRINTABLE_MAX >> 1;
+  std::string value;
+  if (op != AssertOp::DEFAULT) {
+    const StringRef ref = msg.asStrRef();
+    value.append(ref.data(), std::min(ref.size(), MAX_PRINTABLE));
+  }
+
   switch (op) {
   case AssertOp::DEFAULT:
     break;
   case AssertOp::EQ:
   case AssertOp::MATCH: {
-    std::string value;
-    const StringRef ref = msg.asStrRef();
-    constexpr auto MAX_PRINTABLE = SYS_LIMIT_PRINTABLE_MAX >> 1;
-    value.append(ref.data(), std::min(ref.size(), MAX_PRINTABLE));
     value += "\nbinary expression `<LHS> ";
     value += op == AssertOp::EQ ? "==" : "=~";
     value += " <RHS>' is false\n";
@@ -136,6 +139,17 @@ void raiseAssertFail(ARState &st, Value &&msg, const AssertOp op, Value &&left, 
     value += " = ";
     appendAsPrintable(right.hasStrRef() ? right.asStrRef() : right.toString(),
                       MAX_PRINTABLE + value.size(), value);
+    msg = Value::createStr(std::move(value));
+    break;
+  }
+  case AssertOp::IS: {
+    auto &exprType = st.typePool.get(left.getTypeID());
+    auto &targetType = st.typePool.get(right.getTypeID());
+    value += "\nbinary expression `<EXPR> is <TYPE>' is false\n";
+    value += "  <EXPR>: ";
+    appendAsPrintable(exprType.getNameRef(), MAX_PRINTABLE + value.size(), value);
+    value += "\n  <TYPE>: ";
+    appendAsPrintable(targetType.getNameRef(), MAX_PRINTABLE + value.size(), value);
     msg = Value::createStr(std::move(value));
     break;
   }